Add files via upload
Along with four other members, I was able to help create, Maze Runner, a Python-based maze-solving game built with Pygame. The game allows players to navigate through a dynamically generated maze using keyboard controls, aiming to reach the exit before time elapses. It demonstrates the use of pathfinding algorithms, collision detection, and event handling within a game loop. My contributions to the game were developing Win/Lose Conditions,  A Pause Menu (Which was unfortunately not able to be implemented in the final game, due to improper functioning), an Exit menu, and BGM + Sound Effects. Out of 24 groups and 115 students, Our group was selected to present our game to the public to be played with and enjoyed.